requirements.txt
# Core ML/AI Libraries (Compatible versions)
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2

# Computer Vision & Image Processing  
opencv-python==4.8.1.78
Pillow==10.0.1
scikit-image==0.21.0
imageio==2.31.6
matplotlib==3.7.5

# Diffusion Models (Compatible with torch 2.1.2)
diffusers==0.25.1
transformers==4.36.2
accelerate==0.26.1
xformers==0.0.23.post1

# Face Analysis & Processing
insightface==0.7.3
onnxruntime-gpu==1.16.3

# InstantID & ControlNet
huggingface-hub==0.20.3
peft==0.8.2
controlnet-aux==0.3.1

# CodeFormer Dependencies (Custom BasicSR)
lpips==0.1.4
addict==2.4.0
lmdb==1.4.1
yapf==0.40.2

# Core Python Libraries
numpy==1.24.4
scipy==1.11.4
tqdm==4.66.1
requests==2.31.0
gdown==4.7.1

# UI Framework
gradio==4.8.0
fastapi==0.104.1
uvicorn==0.24.0

# Utilities
future==0.18.3
pyyaml==6.0.1
packaging==23.2
typing-extensions==4.8.0

# System dependencies
ninja==1.11.1.1
